added template solution, added R# config, moved MessageContract to separate project, extracted MessageHandler from Worker
