+++ template/.gitignore
@@ -1,0 +1,5 @@
+_ReSharper.*/
+**/bin/*
+**/obj/*
+**/.vs/*
+**.user
