+++ template/Directory.Build.props
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Project>
+   <PropertyGroup>
+      <Version>1.0.0</Version>
+   </PropertyGroup>
+</Project>
