+++ template/src/RabbitMqConsumerTemplate/IMessageHandler.cs
@@ -1,0 +1,9 @@
+namespace RabbitMqConsumerTemplate;
+
+using Byndyusoft.Messaging.RabbitMq;
+using Contracts;
+
+public interface IMessageHandler
+{
+    Task<ConsumeResult> OnMessageReceivedHandler(MessageContract? message, CancellationToken cancellationToken);
+}
