+++ template/src/RabbitMqConsumerTemplate/MessageHandler.cs
@@ -1,0 +1,14 @@
+namespace RabbitMqConsumerTemplate;
+
+using Byndyusoft.Messaging.RabbitMq;
+using Contracts;
+
+public class MessageHandler : IMessageHandler
+{
+    public async Task<ConsumeResult> OnMessageReceivedHandler(MessageContract? message, CancellationToken cancellationToken)
+    {
+        //TODO: implement message processing here
+
+        return ConsumeResult.Ack;
+    }
+}
