+++ template/src/RabbitMqConsumerTemplate/Properties/launchSettings.json
