+++ template/src/RabbitMqConsumerTemplate/RabbitMqConsumerTemplate.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Byndyusoft.Messaging.RabbitMq" Version="0.1.2" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
   </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Contracts\Contracts.csproj" />
+  </ItemGroup>
 </Project>
