+++ template/src/RabbitMqConsumerTemplate/appsettings.Development.json
@@ -4,5 +4,8 @@
             "Default": "Information",
             "Microsoft.Hosting.Lifetime": "Information"
         }
+    },
+    "ConnectionStrings": {
+        "Rabbit": "host=localhost;username=guest;password=guest"
     }
 }
